requirements.txt
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
backcall==0.1.0
cffi==1.12.2
channels==2.1.7
constantly==15.1.0
coverage==4.5.3
cryptography==2.3.1
daphne==2.2.5
decorator==4.3.2
Django==2.1.7
django-cors-headers==2.5.1
django-debug-toolbar==1.11
django-model-utils==3.1.2
django-x509==0.4.1
djangorestframework==3.9.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
jsonfield==2.0.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.3.1
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytz==2018.9
service-identity==18.1.0
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
username==2.0.2
validators==0.12.4
wcwidth==0.1.7
zope.interface==4.6.0
